Make `dbt-labs/facebook-ads` Work With BigQuery & Stitch (#1)
This is a fork of https://github.com/dbt-labs/facebook-ads. As of this writing, the original implementation of this package doesn't support BigQuery. This is a modified version of the original package that works ONLY with BigQuery and Stitch.

This version of the package was tested in the following DBT environment:
- Python 3.8.12
- dbt-bigquery==1.0.0
- dbt-core==1.0.3
- dbt-extractor==0.4.0